requirements.txt
langchain
python-dotenv
pymupdf
langchain-google-genai
pandas